# storage: sync entries to disk in parallel with followers