Referenced in #17500.

This change implements the optimization in the Raft thesis under the
section: 10.2.1 Writing to the leader’s disk in parallel. The optimization
allows the leader to sync new entries to its disk after it has sent the
corresponding `MsgApp` messages, instead of before.

<img width="649" alt="screen shot 2017-08-30 at 3 53 22 pm" src="https://user-images.githubusercontent.com/5438456/29892136-75d391cc-8d9b-11e7-9ff2-6f7aadd7ee46.png">

Here, we invoke this optimization by:
1. sending all MsgApps.
2. syncing all entries and Raft state to disk.
3. sending all other messages.

The write latency speedup we see from this change is promising. On a
write-only workload it's demonstrating anywhere from a **5%-15%**
speedup for average latencies.

For instance, when running:
```
./kv --read-percent=0 -max-block-bytes=512 -min-block-bytes=32
```
we get the following histograms (thanks @jordanlewis!):

#### Zoomed on average latency
![histogram avg](https://user-images.githubusercontent.com/5438456/31524223-20102734-af86-11e7-968e-3c22cd6ace94.png)
#### Zoomed on tail latency
![histogram tail](https://user-images.githubusercontent.com/5438456/31524226-2416c4aa-af86-11e7-923a-16519ac1002e.png)