@nvanbenschoten 2017-10-12
I'd like to do some testing with this on real clusters, but I guess I'll have to wait until
we're out of the cluster drought.