@bdarnell 2017-10-17
> We're a little conservative about when we invoke this optimization, opting
> to only do so when all `Entries` in a `raft.Ready` struct support it. We could
> split the `Entries` into those that need to be synced before the messages are
> sent and those that can be synced after, but that has been seen to exhibit
> worse performance because we end up syncing to disk twice.

The commit message indicates that we decide whether to do this based on the entries, but the code makes the decision based on the messages (writing to disk asynchronously if all messages are MsgApps). I think this is the better approach and could be structured to avoid two calls to sync (instead of "maybe sync, send messages, maybe sync", it could be "maybe send messages, sync, maybe send messages"). 

The latter construction has the nice property that we could easily transform it into "send appends, sync, send other messages". I *think* that would be safe. Keeping MsgApps in order is an important optimization, but I don't think it matters whether we reorder MsgApps relative to other messages. The one thing I'm not sure about is whether it's OK to send out a commit index in MsgApps that we have not ourselves written in HardState.Commit (I assume this is the reason for the IsEmptyHardState check in this diff). I think that's safe - the raft thesis says that the commit index may be treated as volatile state, although etcd/raft keeps it in HardState and I can't remember the details of why. So maybe we keep the IsEmptyHardState check to decide whether we send appends before or after syncing. 

---

Review status: 0 of 1 files reviewed at latest revision, all discussions resolved, all commit checks successful.

---



*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/19229#-:-KwcEsmeuTkZI3ZCAKvy:brhuirh)*
<!-- Sent from Reviewable.io -->