@a-robinson 2017-11-13
How much work would it be to finish this off? I think it'll be a pretty nice win when disks are heavily contended.