@nvanbenschoten 2017-12-01
@bdarnell looking back at the history of `etcd/raft`, I haven't been able to uncover a specific reason why the `Commit` index is included in the `HardState`. It does look like [at one point](https://github.com/coreos/etcd/commit/311db876b017d95dc70e9541000d4b4e7e5d8e1f) it wasn't persisted which may mean that it's just a historical artifact. I'm going to keep digging, but I'd be very interested to hear whether @xiang90 or @yichengq have any insight into this.

Outside of that question, I think I've also been able to convince myself that broadcasting a `MsgApp` with a larger commit index than that in our persisted `HardState.Commit` is still safe. The situation where this gets interesting is when the leader crashes after broadcasting the larger commit index but before persisting it itself. In this case, one of the followers with the entry will necessarily become the next leader, per the leader completeness property and the fact that all majorities intersect. Since only replicas who have seen this entry can become future leaders, the entry will eventually be committed regardless of whether the broadcasted `MsgApp` with the larger commit index reached some of them or not.

Like you mention, Section 3.8 of the Raft thesis states the following:

> Other state variables are safe to lose on a restart, as they can all be recreated. The most interesting example is the commit index, which can safely be reinitialized to zero on a restart. Even if every server restarts at the same time, the commit index will only temporarily lag behind its true value. Once a leader is elected and is able to commit a new entry, its commit index will advance, and it will quickly propagate this commit index to its followers.

So unless I'm missing something there, the only thing that would stop us from unconditionally doing the
> "send appends, sync, send other messages"

approach (regardless of the `IsEmptyHardState` check) would be some implementation detail specific to `etcd/raft`.