@nvanbenschoten 2017-12-05
> The most interesting example is the commit index, which can safely be reinitialized to zero on a restart. 

Thinking about this more, if it isn't necessary to sync the commit index to disk (in both Raft and `etcd/raft`) I wonder if we'd see any difference from clearing it from the `HardState` unconditionally. This would shrink the proto on disk and more importantly avoid the separate key write that we do for about 50% of all messages.