@xiang90 2017-12-06
For etcd, it was just convenient. And we do not really see any performance improvement if we skip commit index sync.