@bdarnell 2017-12-10
<img class="emoji" title=":lgtm:" alt=":lgtm:" align="absmiddle" src="https://reviewable.io/lgtm.png" height="20" width="61"/>

---

Review status: 0 of 1 files reviewed at latest revision, 1 unresolved discussion, all commit checks successful.

---

*[pkg/storage/replica.go, line 3690 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/19229#-L00rPVStbZsD4RSkBN6:-L00rPVStbZsD4RSkBN7:b-lmucxx) ([raw file](https://github.com/cockroachdb/cockroach/blob/a7f0030b9100571f16e99fcd547d8c8e5cabac3b/pkg/storage/replica.go#L3690)):*
> ```Go
> // MsgApps and one containing all other message types. Each slice retains the
> // relative ordering between messages in the original slice.
> func splitMsgApps(msgs []raftpb.Message) (mgsApps, otherMsgs []raftpb.Message) {
> ```

This could use a unittest to make sure we get the boundary conditions right (I'm not sure how common mixed batches are otherwise in our tests)

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/19229#-:-L00vYiGq-qvh2dq8nCN:bnfp4nl)*
<!-- Sent from Reviewable.io -->